+++ src/bosch_shc_camera_client/__init__.py
@@ -1,5 +1,5 @@
 """Async Python client library for Bosch Smart Home Camera cloud + local (RCP) APIs."""
 
-from . import auth_utils, local_rcp
+from . import auth_utils, local_rcp, rcp
 
-__all__ = ["auth_utils", "local_rcp"]
+__all__ = ["auth_utils", "local_rcp", "rcp"]